pyproject.toml
[project]
name = "driver-lisence"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shimizu", email = "reborn.titech@gmail.com" }
]
dependencies = [
    "fastapi[all]>=0.115.6",
    "sqlalchemy>=2.0.36",
    "databases>=0.9.0",
    "pydantic-settings>=2.7.0",
    "jose>=1.0.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "fastapi-debug-toolbar>=0.6.3",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "ulid-py>=1.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/driver_lisence"]
